CA-212333 remove wrapper around standard logging function
The features of the wrapper around Debug.debug and Debug.error is
only used in test/squeeze_test.ml. I have moved the wrapper for
Debug.debug to this file.

The wrapper's main feature is to log to both to syslog and to a file
descriptor (like stdout). As a consequence, now logging in the main
component only goes where the logging facility says it goes. Logging in
squeeze_test.ml goes in addition to stdout as before. "make test", which
runs the test code, still works.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>